# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Meta Service",
    "description": "Wraps the Meta (Facebook) Graph API v22.0 for ad account management, performance reporting, and organic posting.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service key (META_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "AuthorizeResponse": {
        "type": "object",
        "properties": {
          "authorizationUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "authorizationUrl"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "required": [
          "error"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/meta/authorize": {
      "get": {
        "summary": "Generate Meta OAuth login URL",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "required": true,
            "name": "redirectUri",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "label",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth authorization URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/meta/callback": {
      "get": {
        "summary": "Handle Meta OAuth callback",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "code",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "state",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to app with connectionId"
          },
          "400": {
            "description": "Invalid callback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/meta/connections/{connectionId}": {
      "delete": {
        "summary": "Disconnect a Meta connection",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "connectionId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Connection removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}